Adding maximum number of repeated STRs
An additional parameter that should be set by the user. In contrast to nr.STR, which gives you the minimum amount of repetitions, max.nr.STRs gives an upper bound to how many should be searched for. Any repeats longer than that will be discarded. Having the argument set very high (to 150, for example), does not change much in performance, because it does not take long to search for repeats with length of g.e. 140. Usually, the number of repeats with such length is very low.