Updating to poly-repeats
I added the new feature in STR_Detection(). It will search for any STR with at least the repeat length of nr.STR and a maximum repeat length of max.nr.STRs. 
If nchar(STR)==1, it uses the original code. 
Reading into the file and outputting the result remains the same.